remove circular dependency functionality
