+++ .changeset/plenty-yaks-rescue.md
@@ -1,0 +1,5 @@
+---
+'pumpit': major
+---
+
+remove circular dependency functionality
