+++ src/__tests__/child-injector.test.ts
@@ -1,6 +1,5 @@
 import { describe, expect, test } from 'vitest'
 import { PumpIt, SCOPE } from '../pumpit'
-import { get } from '../utils'
 
 describe('Child container', () => {
   test('create child container', () => {
@@ -320,130 +319,6 @@ describe('Child container', () => {
         expect(TestA.count).toBe(2)
         expect(TestB.count).toBe(1)
         expect(childB.keyA).not.toBe(parentA)
-      })
-    })
-
-    describe('Circular injection', () => {
-      test('singleton', () => {
-        const parent = new PumpIt()
-        const child = parent.child()
-        const keyA = Symbol('keyA')
-        const keyB = Symbol('keyB')
-        const keyC = Symbol('keyC')
-
-        class TestA {
-          static count = 0
-
-          constructor() {
-            TestA.count++
-          }
-        }
-
-        class TestB {
-          static count = 0
-
-          static inject = [keyA, get(keyC, { lazy: true })]
-
-          constructor(
-            public keyA: TestA,
-            public keyC: TestC
-          ) {
-            TestB.count++
-          }
-        }
-
-        class TestC {
-          static count = 0
-
-          static inject = [keyA, get(keyB, { lazy: true })]
-
-          constructor(
-            public keyA: TestA,
-            public keyB: TestB
-          ) {
-            TestC.count++
-          }
-        }
-
-        parent.bindClass(keyA, TestA, { scope: 'SINGLETON' })
-        child.bindClass(keyB, TestB, { scope: 'SINGLETON' })
-        child.bindClass(keyC, TestC, { scope: 'SINGLETON' })
-
-        const childB = child.resolve<TestB>(keyB)
-        child.resolve<TestB>(keyB)
-        child.resolve<TestC>(keyC)
-        child.resolve<TestC>(keyC)
-        const parentA = parent.resolve<TestA>(keyA)
-        parent.resolve<TestA>(keyA)
-
-        expect(TestA.count).toBe(1)
-        expect(TestB.count).toBe(1)
-        expect(TestC.count).toBe(1)
-        expect(childB.keyA).toBe(parentA)
-        expect(childB.keyC.keyA).toBe(parentA)
-      })
-
-      test('singleton three levels deep', () => {
-        const grandParent = new PumpIt()
-        const parent = grandParent.child()
-        const child = parent.child()
-        const keyA = Symbol('keyA')
-        const keyB = Symbol('keyB')
-        const keyC = Symbol('keyC')
-
-        class TestA {
-          static count = 0
-
-          constructor() {
-            TestA.count++
-          }
-        }
-
-        class TestB {
-          static count = 0
-
-          static inject = [keyA, get(keyC, { lazy: true })]
-
-          constructor(
-            public keyA: TestA,
-            public keyC: TestC
-          ) {
-            TestB.count++
-          }
-        }
-
-        class TestC {
-          static count = 0
-
-          static inject = [keyA, get(keyB, { lazy: true })]
-
-          constructor(
-            public keyA: TestA,
-            public keyB: TestB
-          ) {
-            TestC.count++
-          }
-        }
-
-        grandParent.bindClass(keyA, TestA, { scope: 'SINGLETON' })
-        child.bindClass(keyB, TestB, { scope: 'SINGLETON' })
-        child.bindClass(keyC, TestC, { scope: 'SINGLETON' })
-
-        const childB = child.resolve<TestB>(keyB)
-        child.resolve<TestB>(keyB)
-        child.resolve<TestC>(keyC)
-        child.resolve<TestC>(keyC)
-
-        const parentA = parent.resolve<TestA>(keyA)
-        parent.resolve<TestA>(keyA)
-
-        grandParent.resolve<TestA>(keyA)
-
-        expect(TestA.count).toBe(1)
-        expect(TestB.count).toBe(1)
-        expect(TestC.count).toBe(1)
-        expect(childB.keyA).toBe(parentA)
-        expect(childB.keyC.keyA).toBe(parentA)
       })
     })
   })
