+++ src/index.ts
@@ -1,5 +1,5 @@
 export * from './pumpit'
 export * from './types'
 
-import { get, getArray, isProxy, transform } from './utils'
-export { get, getArray, transform, isProxy }
+import { get, getArray, transform } from './utils'
+export { get, getArray, transform }
